requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.72
langgraph==0.6.1
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.1
langgraph-sdk==0.2.0
langsmith==0.4.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
openai==1.97.1
openpyxl==3.1.5
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.31
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
zstandard==0.23.0
